+++ RandomSkunk.Results/Result{T}.cs
@@ -45,6 +45,11 @@ public readonly struct Result<T> : IResult<T>, IEquatable<Result<T>>
     public bool IsFail => _outcome == Outcome.Fail;
 
     /// <summary>
+    /// Gets a value indicating whether this is an uninitialized instance of the <see cref="Result{T}"/> struct.
+    /// </summary>
+    public bool IsUninitialized => _outcome == Outcome.Fail && _error is null;
+
+    /// <summary>
     /// Gets the value from the <c>Success</c> result.
     /// </summary>
     /// <returns>If this is a <c>Success</c> result, its value; otherwise throws an <see cref="InvalidStateException"/>.